tools/requirements.txt
# Python dependencies for CRSF utilities

pyserial>=3.5
flask>=2.0.0

